backend/requirements.txt
fastapi
uvicorn
pydantic
weaviate-client
motor
openai
loguru
#transformers
#pytorch
